+++ macros/tmfxEdit.js
@@ -11,19 +11,26 @@ if (!filters.length) return;
 
 let params = filters.map((f) => {
   const tmParams = deepClone(f.tmFilters.tmParams);
-  ['placeableId', 'placeableType', 'filterInternalId', 'filterOwner', 'updateId'].forEach((k) => delete tmParams[k]);
+  ['placeableId', 'placeableType', 'filterInternalId', 'filterOwner', 'updateId'].forEach(
+    (k) => delete tmParams[k]
+  );
   return tmParams;
 });
 if (!params.length) return;
 
 async function savePreset() {
   let content = `<label>Macro</label>
-  <textarea style="width:100%; height: 300px;" readonly>let params = ${JSON.stringify(params, null, 2)};
+  <textarea style="width:100%; height: 300px;" readonly>let params = ${JSON.stringify(
+    params,
+    null,
+    2
+  )};
 
 await TokenMagic.addUpdateFiltersOnSelected(params);</textarea>
   <label>Preset Name</label><input class="presetName" type="text" value="${
     params[0].filterId ?? params[0].filterType
   }"/>
+  <hr><label>Library</label> <select class="library"><option value="tmfx-main">MAIN</option><option value="tmfx-template">TEMPLATE</option></select>
   `;
   new Dialog({
     title: `Params`,
@@ -32,11 +39,12 @@ await TokenMagic.addUpdateFiltersOnSelected(params);</textarea>
       save: {
         label: 'Save As Preset',
         callback: async (html) => {
-          const presetName = html.find('.presetName').val();
-          if (TokenMagic.getPreset(presetName)) {
-            TokenMagic.deletePreset(presetName);
+          const name = html.find('.presetName').val();
+          const library = html.find('.library').val();
+          if (TokenMagic.getPreset({ name, library })) {
+            TokenMagic.deletePreset({ name, library });
           }
-          TokenMagic.addPreset(presetName, params);
+          TokenMagic.addPreset({ name, library }, params);
         },
       },
     },